ch - Pulling Deacon's net charging changes

+++ Server/SHumanEntity.hpp
@@ -42,5 +42,5 @@ private:
 	bool _isSlipImmune = false;
 	bool _isCharging = false;
 	bool _isSwinging = false;
-	long stuntDuration = 0;
+	float stuntDuration = 0;
 };
